get render texture saving to file
a few things wrong: the position is offset, its cause Harvestable is not pos 0,0 its in the center of the screen. can't change that because it'd move the sprite in the scene too, the render textures are basically latebinding so even if you end() and then alter the sprite, the RT uses the current sprites statuses when it renders sometime later

the other thing wrong is that the sprite is not made invisible, same problem as before, if its invisible in the scene, it'll be invisible in the RT and temporarily setting it to invisible wont stick.

i think the answer is to create a duplicate harvestable and draw that instead.

TODO make the shattersprite use the RT's texture and calc out the TextureRect