only add the blood consumer if the arena has been unlocked (NOTE: buggy)
the way the nuitems are created, it goes over all the configs and assumes that they aren't going to change, so it uses the index of the config as a tag for the node. this means that when you add the nuitem halfway through the apps lifecycle it already has a tag for the say 2nd index so it skips it, and then adds the last one twice.

this'll be okay for now but eventually i'll want to have it so you don't need to restart the app to get the consumer. maybe I can trash the nuitems when you level up a building or something